+++ Sources/FWObjC/Plugin/Preview/FWZoomImageView.h
@@ -128,6 +128,9 @@ NS_SWIFT_NAME(ZoomImageView)
 // 播放 video 时屏幕左上角的关闭按钮中心句柄，默认同导航栏关闭按钮
 @property(nonatomic, copy, nullable) CGPoint (^videoCloseButtonCenter)(void);
 
+// 是否隐藏进度视图，默认NO
+@property(nonatomic, assign) BOOL hidesProgressView;
+
 // 进度视图，居中显示
 @property(nonatomic, strong) UIView<FWProgressViewPlugin> *progressView;
 
@@ -136,6 +139,9 @@ NS_SWIFT_NAME(ZoomImageView)
 
 /// 是否正在播放视频
 @property(nonatomic, assign, readonly) BOOL isPlayingVideo;
+
+/// 是否忽略本地图片缓存，默认NO
+@property(nonatomic, assign) BOOL ignoreImageCache;
 
 /// 开始视频播放
 - (void)playVideo;